Problema resolvido e melhorias

+++ backend/routes/eventos.js
@@ -1,19 +1,18 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const bodyParser = require('body-parser');
-const Eventos = require('../models/eventos');
-const Ingressos = require('../models/ingressos');
+const bodyParser = require("body-parser");
+const Eventos = require("../models/eventos");
+const Ingressos = require("../models/ingressos");
 
-const authenticate = require('../authenticate');
+const authenticate = require("../authenticate");
 
 router.use(bodyParser.json());
 
-router.route('/')
+router.route("/")
   .get(authenticate.verifyUser, async (req, res, next) => {
-    console.log(req.user);
     try {
       const eventos = await Eventos.find({});
-      res.setHeader('Content-Type', 'application/json');
+      res.setHeader("Content-Type", "application/json");
       res.status(200).json(eventos);
     } catch (err) {
       res.statusCode = 404;
@@ -22,32 +21,42 @@ router.route('/')
   })
   .post(authenticate.verifyUser, async (req, res, next) => {
     const eventoExists = await Eventos.exists({ nome: req.body.nome });
+    const userIsEmpresa = req.user.tipo === "empresa";
 
     try {
-      if (eventoExists) {
-        throw "Não é possível existir dois eventos com o mesmo nome."
+      if (userIsEmpresa) {
+        if (eventoExists) {
+          throw "Não é possível existir dois eventos com o mesmo nome."
+        } else {
+          const evento = await Eventos.create(req.body);
+          res.setHeader("Content-Type", "application/json");
+          res.status(200).json(evento);
+        }
       } else {
-        const evento = await Eventos.create(req.body);
-        res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(evento);
+        throw "Somente empresas podem criar eventos."
       }
     } catch (err) {
-      res.status(400).send({ error: err });
+      res.status(!userIsEmpresa ? 401 : 400).send({ error: err });
       next(err);
     }
   })
 
-router.route('/:id')
+router.route("/:id")
   .delete(authenticate.verifyUser, async (req, res, next) => {
     const ingressoExists = await Ingressos.exists({ eventoId: req.params.id });
+    const userIsEmpresa = req.user.tipo === "empresa";
 
     try {
-      if (ingressoExists) {
-        throw "Não é possível excluir evento com ingressos associados."
+      if (userIsEmpresa) {
+        if (ingressoExists) {
+          throw "Não é possível excluir evento com ingressos associados."
+        } else {
+          const evento = await Eventos.deleteOne({ _id: req.params.id });
+          res.setHeader("Content-Type", "application/json");
+          res.status(200).json(evento);
+        }
       } else {
-        const evento = await Eventos.deleteOne({ _id: req.params.id });
-        res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(evento);
+        throw "Somente empresas podem deletar eventos."
       }
     } catch (err) {
       res.status(400).send({ error: err });
@@ -57,14 +66,19 @@ router.route('/:id')
   .put(authenticate.verifyUser, async (req, res, next) => {
     const ingressoExists = await Ingressos.exists({ eventoId: req.params.id });
     const eventoEditado = await Eventos.find({ _id: req.params.id });
+    const userIsEmpresa = req.user.tipo === "empresa";
 
     try {
-      if ((eventoEditado[0].nome != req.body.nome) && ingressoExists) {
-        throw "Não é possível editar nome do evento com ingressos associados."
+      if (userIsEmpresa) {
+        if ((eventoEditado[0].nome != req.body.nome) && ingressoExists) {
+          throw "Não é possível editar nome do evento com ingressos associados."
+        } else {
+          await Eventos.updateOne({ _id: req.params.id }, req.body);
+          res.setHeader("Content-Type", "application/json");
+          res.status(200).json(req.body);
+        }
       } else {
-        await Eventos.updateOne({ _id: req.params.id }, req.body);
-        res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(req.body);
+        throw "Somente empresas podem editar eventos."
       }
     } catch (err) {
       res.status(400).send({ error: err })
